requirements/testing.txt
-r base.txt

selenium==4.10.0
factory_boy==3.2.1